Updating .gitignore. Adding SVProgressHUD using carthage. Implementing userdefaults to persist when the spinner needs to be show up.
