+++ Cartfile
@@ -1,0 +1,1 @@
+github "SVProgressHUD/SVProgressHUD"
