+++ Cartfile.resolved
@@ -1,0 +1,1 @@
+github "SVProgressHUD/SVProgressHUD" "2.2.5"
